requirements.txt
streamlit
pandas
spacy>=3.0
groq
python-dotenv
numpy